optimize quntity inc and dec and delete item

+++ Frontend/src/components/Cart/Trash.jsx
@@ -1,28 +1,43 @@
 import { Trash2 } from 'lucide-react'
 import { useCart } from '../../context/CartContext'
-import { useUser } from '../../context/UserContext'
 import api from '../../api/axios'
+import { useState } from 'react'
+
 const Trash = ({ id }) => {
-  const { fetchCart } = useCart()
-  const { fetchUser } = useUser()
+  const { cart, setCart } = useCart()
+  const [loading, setLoading] = useState(false)
+
   const handleDelete = async id => {
-    console.log('Product Id By trash: ', id)
+    if (loading) return
+
+    const previousCart = cart
+
+    // UI SE TURANT REMOVE
+    setCart(prev => prev.filter(item => item.product?._id !== id))
+
     try {
-      const response = await api.delete(`/api/cart/removefromcart/${id}`)
-      if (response.status === 200) {
-        await fetchUser()
-        await fetchCart()
-      }
+      setLoading(true)
+
+      // API CALL
+      await api.delete(`/api/cart/removefromcart/${id}`)
     } catch (error) {
-      console.log(error)
+      console.log('Remove cart error:', error)
+
+      // API FAIL → ITEM WAPAS
+      setCart(previousCart)
+    } finally {
+      setLoading(false)
     }
   }
+
   return (
     <div>
       <button
+        disabled={loading}
         onClick={() => handleDelete(id)}
         className='text-red-400 hover:text-red-500
-        cursor-pointer transition'
+        cursor-pointer transition
+        disabled:cursor-not-allowed'
       >
         <Trash2 size={19} />
       </button>
